Revert "refactor(fixtures): move some fixutres to global conftest"
This reverts commit 79a5ce4c53e07153d97b6c2de405c087fc5aee0b.
